Reorganize new style client tools.
