+++ rdr_service/client/client_libs/README.md
